全面修复时间戳数据类型溢出问题
- 修复SensorFrame结构体：添加rawTimestamp字段(uint64_t)存储原始时间戳
- 修复CommandHandler::formatTimestamp方法参数类型为uint64_t
- 修复WebSocketClient使用rawTimestamp避免精度丢失
- 修复UartReceiver中时间戳处理逻辑
- 更新文档中的方法签名
- 解决所有涉及时间戳计算的数据类型溢出问题

+++ include/CommandHandler.h
@@ -116,7 +116,7 @@ private:
     void showUartConfig(const String& args = "");
     
     // 格式化时间戳
-    String formatTimestamp(uint32_t timestamp);
+    String formatTimestamp(uint64_t timestamp);
     
     // 格式化浮点数
     String formatFloat(float value, int precision = 2);
